Moves code to display.py
- additionaly imports matplot at module level
